refactored Visualizsation and DataUtilities class
